Intelligent version of the script

+++ C Files/Exercises/set_pink/myScript.sh
@@ -1,5 +1,15 @@
-rm exe
-rm *.o
+#!/bin/bash
+
+files=$(find . -name "*.o")
+
+if [ -f "exe" ]; then
+	rm exe
+fi
+
+if [ -n "$files" ]; then
+	rm *.o
+fi
+
 gcc -c grocerystore.c && gcc -c main.c
 gcc -o exe grocerystore.o main.o
 ./exe
